+++ opentelemetry-etw-logs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## vNext
 
+## v0.11.1
+- Do not panic on debug builds when transient errors are returned by `EventBuilder::write`.
+  [#737](https://github.com/open-telemetry/opentelemetry-rust-contrib/pull/737)
+
 ## v0.11.0
 
 - Bump opentelemetry and opentelemetry_sdk versions to 0.32